+++ sdk/storagemover/azure-resourcemanager-storagemover/src/main/java/com/azure/resourcemanager/storagemover/implementation/JobDefinitionImpl.java
@@ -18,6 +18,7 @@ import com.azure.resourcemanager.storagemover.models.JobRunStatus;
 import com.azure.resourcemanager.storagemover.models.JobType;
 import com.azure.resourcemanager.storagemover.models.ProvisioningState;
 import com.azure.resourcemanager.storagemover.models.ScheduleInfo;
+import java.time.OffsetDateTime;
 import java.util.Collections;
 import java.util.List;
 
@@ -125,6 +126,26 @@ public final class JobDefinitionImpl implements JobDefinition, JobDefinition.Def
 
     public Boolean preservePermissions() {
         return this.innerModel().preservePermissions();
+    }
+
+    public Boolean isCrossTenantJob() {
+        return this.innerModel().isCrossTenantJob();
+    }
+
+    public String crossTenantEndpointTenantId() {
+        return this.innerModel().crossTenantEndpointTenantId();
+    }
+
+    public String crossTenantEndpointResourceId() {
+        return this.innerModel().crossTenantEndpointResourceId();
+    }
+
+    public String syncMode() {
+        return this.innerModel().syncMode();
+    }
+
+    public OffsetDateTime moverSyncedUntil() {
+        return this.innerModel().moverSyncedUntil();
     }
 
     public String resourceGroupName() {
@@ -250,6 +271,16 @@ public final class JobDefinitionImpl implements JobDefinition, JobDefinition.Def
             .stopJob(resourceGroupName, storageMoverName, projectName, jobDefinitionName);
     }
 
+    public Response<JobRunResourceId> reconcileJobWithResponse(Context context) {
+        return serviceManager.jobDefinitions()
+            .reconcileJobWithResponse(resourceGroupName, storageMoverName, projectName, jobDefinitionName, context);
+    }
+
+    public JobRunResourceId reconcileJob() {
+        return serviceManager.jobDefinitions()
+            .reconcileJob(resourceGroupName, storageMoverName, projectName, jobDefinitionName);
+    }
+
     public JobDefinitionImpl withCopyMode(CopyMode copyMode) {
         if (isInCreateMode()) {
             this.innerModel().withCopyMode(copyMode);
@@ -343,6 +374,41 @@ public final class JobDefinitionImpl implements JobDefinition, JobDefinition.Def
     public JobDefinitionImpl withPreservePermissions(Boolean preservePermissions) {
         this.innerModel().withPreservePermissions(preservePermissions);
         return this;
+    }
+
+    public JobDefinitionImpl withIsCrossTenantJob(Boolean isCrossTenantJob) {
+        this.innerModel().withIsCrossTenantJob(isCrossTenantJob);
+        return this;
+    }
+
+    public JobDefinitionImpl withCrossTenantEndpointTenantId(String crossTenantEndpointTenantId) {
+        this.innerModel().withCrossTenantEndpointTenantId(crossTenantEndpointTenantId);
+        return this;
+    }
+
+    public JobDefinitionImpl withCrossTenantEndpointResourceId(String crossTenantEndpointResourceId) {
+        this.innerModel().withCrossTenantEndpointResourceId(crossTenantEndpointResourceId);
+        return this;
+    }
+
+    public JobDefinitionImpl withSyncMode(String syncMode) {
+        if (isInCreateMode()) {
+            this.innerModel().withSyncMode(syncMode);
+            return this;
+        } else {
+            this.updateJobDefinition.withSyncMode(syncMode);
+            return this;
+        }
+    }
+
+    public JobDefinitionImpl withMoverSyncedUntil(OffsetDateTime moverSyncedUntil) {
+        if (isInCreateMode()) {
+            this.innerModel().withMoverSyncedUntil(moverSyncedUntil);
+            return this;
+        } else {
+            this.updateJobDefinition.withMoverSyncedUntil(moverSyncedUntil);
+            return this;
+        }
     }
 
     private boolean isInCreateMode() {